:zap: Improve CAPTCHA performance.
Verifying a CAPTCHA requires
an HTTP request to the CAPTCHA platform,
which until then was synchronous.
This commit makes it an asynchronous request.
This improves performance.

:white_check_mark: Update test cases.

This commit introduces a thread pool,
but the name and parameters of the thread pool
must be configured in the configuration file before
using the thread pool, so update the test cases.